added a screen to vr for testing
