+++ src/navigation/AppNavigator.js
@@ -1,0 +1,2 @@
+import React from 'react';
+import {}
